+++ tests/unit/date-range-input.spec.ts
@@ -220,7 +220,12 @@ describe('Date range input', () => {
     });
 
     describe('when clicking on save button', () => {
-      const editedValue: RelativeDateRange = { date: '{{today}}', quantity: -1, duration: 'month' };
+      const editedValue: RelativeDateRange = {
+        quantity: 1,
+        duration: 'month',
+        operator: 'until',
+        date: '{{today}}',
+      };
 
       beforeEach(async () => {
         wrapper.find('RelativeDateRangeForm-stub').vm.$emit('input', editedValue);
@@ -262,7 +267,12 @@ describe('Date range input', () => {
     });
 
     describe('choose right selected tab when opening calendar', () => {
-      const initialValue: RelativeDateRange = { date: '{{today}}', quantity: 1, duration: 'month' };
+      const initialValue: RelativeDateRange = {
+        date: '{{today}}',
+        quantity: 1,
+        duration: 'month',
+        operator: 'from',
+      };
       const updatedValue = { start: new Date(), end: new Date(1) };
       beforeEach(async () => {
         createWrapper({
@@ -295,7 +305,7 @@ describe('Date range input', () => {
       });
 
       describe('when updating RelativeDateRangeForm value', () => {
-        const newValue = { date: '{{today}}', quantity: -1, duration: 'month' };
+        const newValue = { quantity: 1, duration: 'month', operator: 'until', date: '{{today}}' };
         beforeEach(async () => {
           wrapper.find('RelativeDateRangeForm-stub').vm.$emit('input', newValue);
           await wrapper.vm.$nextTick();
@@ -310,7 +320,12 @@ describe('Date range input', () => {
     });
 
     describe('when switching between tabs', () => {
-      const updatedRelativeDateValue = { date: '{{today}}', quantity: -1, duration: 'month' };
+      const updatedRelativeDateValue = {
+        quantity: 1,
+        duration: 'month',
+        operator: 'until',
+        date: '{{today}}',
+      };
       beforeEach(async () => {
         wrapper.find('RelativeDateRangeForm-stub').vm.$emit('input', updatedRelativeDateValue); // update RelativeDateRangeForm value
         await wrapper.vm.$nextTick();
@@ -390,7 +405,12 @@ describe('Date range input', () => {
   });
 
   describe('with selected value as relative date', () => {
-    const value: RelativeDateRange = { date: '{{today}}', quantity: 1, duration: 'month' };
+    const value: RelativeDateRange = {
+      date: '{{today}}',
+      quantity: 1,
+      duration: 'month',
+      operator: 'from',
+    };
     beforeEach(() => {
       createWrapper({
         availableVariables: SAMPLE_VARIABLES,
@@ -498,7 +518,12 @@ describe('Date range input', () => {
   });
 
   describe('with bounds', () => {
-    const bounds: RelativeDateRange = { date: '{{today}}', quantity: 1, duration: 'month' };
+    const bounds: RelativeDateRange = {
+      date: '{{today}}',
+      quantity: 1,
+      duration: 'month',
+      operator: 'from',
+    };
     beforeEach(() => {
       createWrapper({
         availableVariables: SAMPLE_VARIABLES,